chore: standardize documentation comments
Those two changes aim to be more respectful of PHPDoc, resulting in a
better integration of, for instance, PhpStorm's documentation comments
rendering which gives a nice integration of code blocks, links, etc. for
users.

Changing "```php" to "```", because the first one breaks the code block
rendering.

Changing "@see some-link" to "{@see some-link}".